# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Restaurant Orders",
    "description" : "Simple API to manage restaurant orders",
    "contact" : {
      "name" : "Sergio Adonis",
      "email" : "sergio.adonis@outlook.com"
    },
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/sergioadonis/restaurant-orders-api/v1",
    "description" : "SwaggerHub API Auto Mocking"
  } ],
  "paths" : {
    "/categories" : {
      "get" : {
        "tags" : [ "Category" ],
        "operationId" : "getCategories",
        "parameters" : [ {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "default" : 50
          }
        }, {
          "name" : "skip",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "default" : 0
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CategoryArrayResult"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Category" ],
        "operationId" : "addCategory",
        "requestBody" : {
          "$ref" : "#/components/requestBodies/categoryBody"
        },
        "responses" : {
          "default" : {
            "description" : "ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CategoryObjectResult"
                }
              }
            }
          }
        }
      }
    },
    "/categories/{id}" : {
      "get" : {
        "tags" : [ "Category" ],
        "operationId" : "getCategoryById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/CommonProps/properties/id"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CategoryObjectResult"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Category" ],
        "operationId" : "updateCategory",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/CommonProps/properties/id"
          }
        } ],
        "requestBody" : {
          "$ref" : "#/components/requestBodies/categoryBody"
        },
        "responses" : {
          "default" : {
            "description" : "ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CategoryObjectResult"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Category" ],
        "operationId" : "deleteCategory",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/CommonProps/properties/id"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CategoryObjectResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Category" : {
        "type" : "object",
        "allOf" : [ {
          "$ref" : "#/components/schemas/CommonProps"
        } ]
      },
...
      "CommonProps" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64",
            "readOnly" : true,
            "example" : 123
          },
          "name" : {
            "type" : "string",
            "example" : "Name of object"
          },
          "description" : {
            "type" : "string",
            "example" : "This is a description for this object, could be Markdown sintax."
          },
          "photoUrl" : {
            "type" : "string",
            "example" : "https://user-contents.domain.example.com/12ea34-651d76c-87bd-85b6f9"
          }
        },
        "example" : {
          "photoUrl" : "https://user-contents.domain.example.com/12ea34-651d76c-87bd-85b6f9",
          "name" : "Name of object",
          "description" : "This is a description for this object, could be Markdown sintax.",
          "id" : 123
        }
      },
...
      "Result" : {
        "type" : "object",
        "properties" : {
          "errorCode" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "requestId" : {
            "type" : "string"
          }
        }
      },
      "CategoryArrayResult" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/CategoryArrayResult_data"
          }
        },
        "example" : {
          "data" : {
            "categories" : [ "", "" ]
          }
        },
        "allOf" : [ {
          "$ref" : "#/components/schemas/Result"
        } ]
      },
      "CategoryObjectResult" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/CategoryObjectResult_data"
          }
        },
        "example" : {
          "data" : {
            "category" : {
              "restaurants" : [ {
                "photoUrl" : "https://user-contents.domain.example.com/12ea34-651d76c-87bd-85b6f9",
                "name" : "Name of object",
                "description" : "This is a description for this object, could be Markdown sintax.",
                "id" : 123
              }, {
                "photoUrl" : "https://user-contents.domain.example.com/12ea34-651d76c-87bd-85b6f9",
                "name" : "Name of object",
                "description" : "This is a description for this object, could be Markdown sintax.",
                "id" : 123
              } ]
            }
          }
        },
        "allOf" : [ {
          "$ref" : "#/components/schemas/Result"
        } ]
      },
...
      "CategoryArrayResult_data" : {
        "type" : "object",
        "properties" : {
          "categories" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Category"
            }
          }
        },
        "example" : {
          "categories" : [ "", "" ]
        }
      },
      "CategoryObjectResult_data_category" : {
        "type" : "object",
        "properties" : {
          "restaurants" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CommonProps"
            }
          }
        },
        "example" : {
          "restaurants" : [ {
            "photoUrl" : "https://user-contents.domain.example.com/12ea34-651d76c-87bd-85b6f9",
            "name" : "Name of object",
            "description" : "This is a description for this object, could be Markdown sintax.",
            "id" : 123
          }, {
            "photoUrl" : "https://user-contents.domain.example.com/12ea34-651d76c-87bd-85b6f9",
            "name" : "Name of object",
            "description" : "This is a description for this object, could be Markdown sintax.",
            "id" : 123
          } ]
        }
      },
      "CategoryObjectResult_data" : {
        "type" : "object",
        "properties" : {
          "category" : {
            "$ref" : "#/components/schemas/CategoryObjectResult_data_category"
          }
        },
        "example" : {
          "category" : {
            "restaurants" : [ {
              "photoUrl" : "https://user-contents.domain.example.com/12ea34-651d76c-87bd-85b6f9",
              "name" : "Name of object",
              "description" : "This is a description for this object, could be Markdown sintax.",
              "id" : 123
            }, {
              "photoUrl" : "https://user-contents.domain.example.com/12ea34-651d76c-87bd-85b6f9",
              "name" : "Name of object",
              "description" : "This is a description for this object, could be Markdown sintax.",
              "id" : 123
            } ]
          }
        }
      },
...
    },
...
    "requestBodies" : {
      "categoryBody" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Category"
            }
          }
        },
        "required" : true
      },
...
    }
  }
}